add post functionality
